refactor(tests): clean up
