+++ tests/conftest.py
@@ -14,11 +14,10 @@ except ImportError:
 import requests
 import pytest
 import sqlalchemy as sa
-from sqlalchemy.engine.url import URL
 
 
 from rs_sqla_test_utils import db
-from packaging.version import Version
+from rs_sqla_test_utils.utils import make_mock_engine
 
 
 _unicode = type(u'')
@@ -47,7 +46,9 @@ class DatabaseTool(object):
 
     @contextlib.contextmanager
     def _database(self):
-        from sqlalchemy_redshift.dialect import RedshiftDialect_redshift_connector
+        from sqlalchemy_redshift.dialect import (
+            RedshiftDialect_redshift_connector
+        )
 
         db_name = database_name()
         with self.engine.connect() as conn:
@@ -70,7 +71,9 @@ class DatabaseTool(object):
         finally:
             with self.engine.connect() as conn:
                 conn.execute('COMMIT')  # Can't drop databases in a transaction
-                if isinstance(conn.dialect, RedshiftDialect_redshift_connector):
+                if isinstance(
+                        conn.dialect, RedshiftDialect_redshift_connector
+                ):
                     conn.execution_options(isolation_level="AUTOCOMMIT")
                 conn.execute('DROP DATABASE {db_name}'.format(db_name=db_name))
 
@@ -101,85 +104,39 @@ def pytest_addoption(parser):
     parser.addoption("--dbdriver", action="append")
 
 
-def make_mock_engine(name):
-    """
-    Creates a mock sqlalchemy engine for testing dialect functionality
-
-    """
-    if Version(sa.__version__) >= Version('1.4.0'):
-        return sa.create_mock_engine(URL(
-            drivername=name
-        ), executor=None)
-    else:
-        return sa.create_engine(URL(
-            drivername=name,
-        ), strategy='mock', executor=None)
-
-
 class DriverParameterizedTests:
     """
     Helper class for generating fixture params using pytest config opts.
 
     """
     DEFAULT_DRIVERS = ['psycopg2', 'psycopg2cffi']
-    drivers = None
-    stub_redshift_engines = None
-    stub_redshift_dialects = None
     redshift_dialect_flavors = None
 
     @classmethod
     def set_drivers(cls,  _drivers):
-        DriverParameterizedTests.drivers = _drivers
-        DriverParameterizedTests.make_fixtures()
-
-    @staticmethod
-    def _make_drivername(driver):
-        return 'redshift+{}'.format(driver)
-
-    @classmethod
-    def make_fixtures(cls):
-
-        driver_names = [
-            DriverParameterizedTests._make_drivername(x)
-            for x in DriverParameterizedTests.drivers
+        DriverParameterizedTests.redshift_dialect_flavors = [
+            'redshift+{}'.format(x) for x in _drivers
         ]
-
-        DriverParameterizedTests.stub_redshift_engines = [
-                make_mock_engine(x)
-                for x in driver_names
-        ]
-
-        DriverParameterizedTests.stub_redshift_dialects = [
-            x.dialect for x in DriverParameterizedTests.stub_redshift_engines
-        ]
-
-        DriverParameterizedTests.redshift_dialect_flavors = driver_names
 
 
 def pytest_generate_tests(metafunc):
-    if DriverParameterizedTests.drivers is None:
-        dbdrivers = metafunc.config.getoption("--dbdriver", default=DriverParameterizedTests.DEFAULT_DRIVERS)
-        DriverParameterizedTests.set_drivers(dbdrivers)
 
-    if 'stub_redshift_engine' in metafunc.fixturenames:
-        metafunc.parametrize(
-            "stub_redshift_engine",
-            DriverParameterizedTests.stub_redshift_engines,
-            scope="session")
-    if 'stub_redshift_dialect' in metafunc.fixturenames:
-        metafunc.parametrize(
-            'stub_redshift_dialect',
-            DriverParameterizedTests.stub_redshift_dialects,
-            scope="session")
     if 'redshift_dialect_flavor' in metafunc.fixturenames:
+        if DriverParameterizedTests.redshift_dialect_flavors is None:
+            dbdrivers = metafunc.config.getoption(
+                "--dbdriver", default=DriverParameterizedTests.DEFAULT_DRIVERS
+            )
+            DriverParameterizedTests.set_drivers(dbdrivers)
+
         metafunc.parametrize(
             'redshift_dialect_flavor',
             DriverParameterizedTests.redshift_dialect_flavors,
+            ids=DriverParameterizedTests.redshift_dialect_flavors,
             scope="session")
 
 
 @pytest.yield_fixture(scope='session')
-def _redshift_database_tool(request, redshift_dialect_flavor):
+def _redshift_database_tool(redshift_dialect_flavor):
     from rs_sqla_test_utils import models
     if 'PGPASSWORD' not in os.environ:
         pytest.skip('This test will only work on Travis.')
@@ -263,3 +220,13 @@ def redshift_session(_session_scoped_redshift_engine):
         session.close()
         tx.rollback()
         conn.close()
+
+
+@pytest.fixture(scope='session')
+def stub_redshift_engine(redshift_dialect_flavor):
+    yield make_mock_engine(redshift_dialect_flavor)
+
+
+@pytest.fixture(scope='session')
+def stub_redshift_dialect(stub_redshift_engine):
+    yield stub_redshift_engine.dialect
