+++ tests/rs_sqla_test_utils/utils.py
@@ -1,6 +1,9 @@
 __author__ = 'haleemur'
 
 import re
+from packaging.version import Version
+import sqlalchemy as sa
+from sqlalchemy.engine.url import URL
 
 
 def clean(query):
@@ -14,3 +17,18 @@ def compile_query(q, _dialect):
         dialect=_dialect,
         compile_kwargs={'literal_binds': True})
     )
+
+
+def make_mock_engine(name):
+    """
+    Creates a mock sqlalchemy engine for testing dialect functionality
+
+    """
+    if Version(sa.__version__) >= Version('1.4.0'):
+        return sa.create_mock_engine(URL(
+            drivername=name
+        ), executor=None)
+    else:
+        return sa.create_engine(URL(
+            drivername=name,
+        ), strategy='mock', executor=None)
