+++ tox.ini
@@ -5,9 +5,14 @@ envlist =
     lint,
     docs
 
-[testenv:py{27,35}-sa{13,14}]
+[testenv]
 passenv = PGPASSWORD
-commands = pytest {posargs} --dbdriver psycopg2 --dbdriver psycopg2cffi
+deps =
+    sa13: sqlalchemy==1.3.24
+    sa14: sqlalchemy==1.4.15
+
+[testenv:py{27,35}-sa{13,14}]
+commands = pytest -s {posargs} --dbdriver psycopg2 --dbdriver psycopg2cffi
 deps =
     alembic==1.4.2
     packaging==20.4
@@ -15,8 +20,6 @@ deps =
     psycopg2cffi==2.8.1
     pytest==3.10.1
     requests==2.7.0
-    sa13: sqlalchemy==1.3.24
-    sa14: sqlalchemy==1.4.15
 
 [testenv:py{36,37,38,39}-sa{13,14}]
 commands = pytest {posargs} --dbdriver psycopg2 --dbdriver psycopg2cffi --dbdriver redshift_connector
